make it work with miriad=.

+++ docs/install_miriad.sh
@@ -3,7 +3,7 @@
 #  new V6.0+ install
 #
 
-echo "install_miriad.sh:  Version 1.0 -- 29-may-2020"
+echo "install_miriad.sh:  Version 1.0 -- 28-jan-2023"
 
 opt=1
 miriad=miriad
@@ -49,11 +49,19 @@ fi
 
 date0=$(date)
 
-rm -rf $miriad
-git clone $url $miriad
-cd $miriad
-git checkout $branch
-
+if [ $miriad == "." ]; then
+    if [ -d .git ]; then
+	git pull
+    else
+	echo Not a miriad with git?
+	exit 0
+    fi
+else
+    rm -rf $miriad
+    git clone $url $miriad
+    cd $miriad
+    git checkout $branch
+fi
 
 #                            default use the GNU compiler
 export F77=gfortran
